Backend/requirements.txt
flask
flask-cors
flask-bcrypt
flask-jwt-extended
psycopg2-binary
tensorflow
pillow
numpy
werkzeug
python-dotenv
gunicorn